Fix reinstall

+++ scripts/0.1.16/postrm.sh
@@ -1,3 +1,6 @@
 #!/usr/bin/env bash
 
-rm -f /usr/bin/hr
+fname=$(readlink -f /usr/bin/hr)
+if [[ $fname == "/opt/hansonrobotics/hrtool/hr-base" ]]; then
+    rm -f /usr/bin/hr
+fi
